--- src/main/java/juja/microservices/gamification/slackbot/controller/AchievementController.java
@@ -1,11 +1,0 @@
-package juja.microservices.gamification.slackbot.controller;
-
-/**
- * Created by Nikol on 3/4/2017.
- */
-
-public class AchievementController {
-    public void sendCodenjoy(String from, String firstPlace, String secondPlace, String thirdPlace){
-
-    }
-}
